requirements.txt
aFlask==2.3.2
Flask-Login==0.6.2
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.9.1
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
eventlet==0.35.1
